fix match_modules_set to work with MoE
Summary:

match_modules_set isn't currently as useful as it could be because it
lacks the ability to match multiple results for each set like in the
case of a moe model where you have 128 experts.

```
[`layers.32.mlp.experts.0.gate_up_proj`, ..., `layers.32.mlp.experts.127.gate_up_proj`]
```

In order to make is so this can still work for matching simple cases and
moe cases we use the following approach.

1) match modules until we have at least 1 match per target
2) when we have 1 match per target, our set is 'full' and we calculate
   the common parent context
3) continue matching and for each match, check if parent context would change given the
   new match
4) if we find a match that changes the parent context, this is the first
   element of the next set. yield the existing matched set and then
reset, using the current match as the first element of the new set.

To facilitate this algorithm i also added get_lowest_common_module_name
which basically copies a similar function in llm-compressor though
significantly simpler.

Signed-off-by: HDCharles <charlesdavidhernandez@gmail.com>